Moves Layer source and header files to /src and /include
