+++ src/Layer.cpp
@@ -1,9 +1,9 @@
 /*! \file
  *
  */
-#include "Layer.h"
+#include "../include/Layer.h"
 
-#include "../../TEMLogger.h"
+#include "TEMLogger.h"
 extern src::severity_logger< severity_level > glg;
 
 Layer::Layer() {
