+++ src/ecodomain/layer/ParentLayer.h
@@ -4,7 +4,7 @@
 #ifndef PARENTLAYER_H_
 #define PARENTLAYER_H_
 
-#include "Layer.h"
+#include "../../../include/Layer.h"
 
 #include <string>
 #include <cmath>
