+++ src/ecodomain/layer/SoilLayer.h
@@ -3,7 +3,7 @@
  */
 #ifndef SOILLAYER_H_
 #define SOILLAYER_H_
-#include "Layer.h"
+#include "../../../include/Layer.h"
 
 #include <math.h>
 #include <cmath>
