+++ src/snowsoil/TemperatureUpdator.h
@@ -11,7 +11,7 @@
 #include <cmath>
 
 #include "../../include/Ground.h"
-#include "../ecodomain/layer/Layer.h"
+#include "../../include/Layer.h"
 #include "../ecodomain/layer/SnowLayer.h"
 #include "../ecodomain/layer/SoilLayer.h"
